+++ database/migrations/2020_03_04_153910_create_kecambahs_table.php
@@ -17,9 +17,9 @@ class CreateKecambahsTable extends Migration
             $table->bigIncrements('kecambah_id');
             $table->string('nama');
             $table->string('pembeli');
-            $table->integer('jumlah');
-            $table->integer('harga');
-            $table->integer('jumlah_bayar');
+            $table->bigInteger('jumlah');
+            $table->bigInteger('harga');
+            $table->bigInteger('jumlah_bayar');
             $table->timestamps();
             $table->softDeletes();
         });
